Move ASP.NET related code into a separate library dotless.AspNet. Change dotless.Core to target .NET 3.5 Client Profile. Added class AspNetContainerFactory that inherits ContainerFactory and registers the ASP.NET specific services.
Add Generate-Assembly-Info to build script for dotless.AspNet project.
Fix msbuild call in default.ps1 to allow spaces in paths.
