+++ src/dotless.AspNet/Input/VirtualFileReader.cs
