+++ src/dotless.AspNet/Loggers/AspNetTraceLogger.cs
