+++ src/dotless.AspNet/Response/CssResponse.cs
@@ -1,6 +1,5 @@
 namespace dotless.Core.Response
 {
-    using System.Web;
     using Abstractions;
 
     public class CssResponse : IResponse
