+++ src/dotless.SampleWeb/dotless.SampleWeb.csproj
@@ -17,6 +17,7 @@
     </FileUpgradeFlags>
     <OldToolsVersion>3.5</OldToolsVersion>
     <UpgradeBackupLocation />
+    <UseIISExpress>false</UseIISExpress>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -129,6 +130,10 @@
     <Content Include="Content\Web.config" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\dotless.AspNet\dotless.AspNet.csproj">
+      <Project>{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}</Project>
+      <Name>dotless.AspNet</Name>
+    </ProjectReference>
     <ProjectReference Include="..\dotless.Core\dotless.Core.csproj">
       <Project>{D7A24E58-68D0-4B10-A324-EA33C822FA56}</Project>
       <Name>dotless.Core</Name>
