+++ src/dotless.Test/dotless.Test.csproj
@@ -82,6 +82,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Config\AspNetConfigurationFixture.cs" />
     <Compile Include="Plugins\ColorSpinPluginFixture.cs" />
     <Compile Include="Plugins\PluginFixture.cs" />
     <Compile Include="Plugins\RtlPluginFixture.cs" />
@@ -154,6 +155,10 @@
     <Compile Include="HttpFixtureBase.cs" />
   </ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\dotless.AspNet\dotless.AspNet.csproj">
+      <Project>{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}</Project>
+      <Name>dotless.AspNet</Name>
+    </ProjectReference>
     <ProjectReference Include="..\dotless.Core\dotless.Core.csproj">
       <Project>{D7A24E58-68D0-4B10-A324-EA33C822FA56}</Project>
       <Name>dotless.Core</Name>
