+++ src/dotless.sln
@@ -18,6 +18,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotless.Benchmark", "dotles
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotless.SampleWeb", "dotless.SampleWeb\dotless.SampleWeb.csproj", "{87DE70C6-FF78-4071-B6B6-710997997B7B}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotless.AspNet", "dotless.AspNet\dotless.AspNet.csproj", "{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -44,6 +46,10 @@ Global
 		{87DE70C6-FF78-4071-B6B6-710997997B7B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{87DE70C6-FF78-4071-B6B6-710997997B7B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{87DE70C6-FF78-4071-B6B6-710997997B7B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{27FDE40F-E671-4C5C-BC24-FD04EA6EA2B2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
